refactor: Process data placeholder for SubmitFormResultRequest
Moved 2 functions from total_mobile_handler.py to parse_json.py

Refs: BLAIS5-3086
