+++ app/handlers/total_mobile_handler.py
@@ -1,5 +1,5 @@
 from app.services.total_mobile_service import do_something_service, update_case_telephone_number
-from app.utilities.parse_json import get_case_details, get_telephone_number
+from app.utilities.parse_json import validate_request, get_case_details, get_telephone_number, get_reference_number
 
 
 def submit_form_result_request_handler(request):
@@ -24,22 +24,3 @@ def complete_visit_request_handler(request):
     data = validate_request(request)
     reference_number = get_reference_number(data)
     do_something_service(reference_number)
-
-
-def validate_request(request):
-    # TODO: Test dis
-    data = request.get_json()
-    if not data:
-        print("Ain't got no data, mate")
-        raise ValueError()
-    print("We gots data!")
-    return data
-
-
-def get_reference_number(data):
-    print("This is completely arbitrary data")
-    try:
-        return data["Identity"]["Reference"]
-    except Exception as err:
-        print(f"Failed to get reference number: {err}")
-        raise err
